[💥✨] Add support for multiple repos to CLI
In organizations with lots of repos, its common to have pull requests
across many of them. Having to run the command multiple times to get
list of pending PRs is tedious and annoying to other members of the
organization.

Breaking changes:
The long flag `--repo` has been renamed to `--repos`. This shouldn't
cause any issues, as the CLI parser I use looks for
shortest-matching-flag